refactor(tools): drop execute_builder_query limit clamp; tidy comments
The 64 MiB doRequest response guard already bounds execute_builder_query, so
clamping the limit embedded in the caller's authored query added surface area
(clampBuilderQueryLimits + builderQueryResult) for only a ~64->16 MiB gain.
Revert the tool to NewToolResultText and remove the clamp, its wrapper, and
its tests. Also trim the verbose comments added across the PR.
